Update testdisk-remove.sh with pwd fix

+++ testdisk-remove.sh
@@ -1,11 +1,22 @@
 #!/bin/bash
 
+current_dir=$(pwd)
+directory=$1
+number=$2
 
-number=$1
+if [ "$directory" != "" ]
+then
 
-for remove_item in $(seq 1 $number)
-do
-        rm -r -f recup_dir.$remove_item
-done
+ cd $directory
 
+ for remove_item in $(seq 1 $number)
+ do
+   rm -r -f recup_dir.$remove_item
+ done
+
+ cd $current_dir
+
+else
+  echo "please pass your test-disk recover directory"
+fi
